Merged PR 63842: Address submission feedback from Adobe
## What's being changed

- Remove logging of headers
- Restore the starter-kit-info runtime action
- Rename ContactPatchHandler > CustomerConsumer
- Remove COMMERCE_ADMIN_TOKEN and related code
- Remove unused vars
- Remove dotenv ref
- Remove configSchema

## Why it's being changed

`configSchema` is removed from app.config.yaml because our app is primarily a downloadable accelerator. To include the option to configure the app via a UI is a) a bit pointless because all it does is set values in a `.env` file that the installer has to customise anyway, and b) possibly confusing because it suggests to the installer that there the app is installable.

## How to review / test this change

- Smoke tests

Related work items: #293696, #293735